refactor(capture-sdk): Move `ErrorType` to the `error` package
It is closer related to the error screen than to the `network` package as it uses the network
error only in a helper method to map to the appropriate user presented error case.

PIA-3325
