+++ capture-sdk/sdk/src/main/java/net/gini/android/capture/network/FailureException.java
@@ -1,5 +1,7 @@
 package net.gini.android.capture.network;
 
+import net.gini.android.capture.error.ErrorType;
+
 public class FailureException extends Exception {
 
     public final ErrorType errorType;
